# docs/api-documentation.yml
openapi: "3.0.0"
info:
  version: "5.0.0"
  title: "FemDevs API Documentation"
  description: |-
    The documentation for the FemDevs API.
    Do note that some routes have permission-based requirements that your user must be qualified for in order to access them.
    Check the description for each route in case you are unsure.

    Permissions Overrides:

    The following permissions will bypass all permissions unless its marked otherwise:
    * `Global::Role.Developer`
    * `Global::Role.Administrator`
    * `Global::Role.Owner`

    It is also always recommended that you append a query parameter of `?type=json` to each request to ensure that all requests are returned as a JSON object
  termsOfService: "https://thefemdevs.com/legal/terms"
  contact:
    name: "FemDevs"
    url: "https://thefemdevs.com"
    email: "support@thefemdevs.com"
  license:
    name: "GNU Affero General Public License v3.0"
    url: "https://www.gnu.org/licenses/agpl-3.0.en.html"
servers:
  - url: "https://thefemdevs.com/api"
    description: "Production Server"
paths:
...
  /mail/send:
    post:
      operationId: "SendEmail"
      description: |-
        Sends an email to the given recipient.
        Returns a success message if successful, otherwise returns an error.

        Permissions Required: Manager Role (Dev, Admin, or Owner)
      summary: "Send an Email"
      parameters:
        - in: "header"
          name: 'x-user'
          required: true
          description: "The Account Credentials used to send the email"
          example: "XQGdt1KLapMtFZqGCdzyLw.xcWtsV5Xb4T-yEE0lHtLyC60JC4"
          schema:
            type: string
        - in: "header"
          name: "x-from"
          required: true
          description: |-
            The information of the sender to be included in the message (sender's name, sender's email address, etc)

            Syntax: `n:NAME;a:ADDRESS;r:REPLYTOADDRESS`
          example: 'n:Example User;a:main@example.com;r:replyto@example.com'
          schema:
            type: string
      requestBody:
        description: "The email information to send"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
            examples:
              Plain-Text:
                value:
                  to: "user@example.com"
                  subject: "Test"
                  text: "This is a test email"
              HTML:
                value:
                  to: "user@example.com"
                  subject: "Test"
                  html: "<h1>This is a test email</h1>"
              Link:
                value:
                  to: "user@example.com"
                  subject: "Test"
                  url: "https://example.com"         
      responses:
        200:
          description: "Ok Response"
          content:
            text/plain:
              schema:
                type: string
              examples:
                Good:
                  value: "OK"
        400:
          description: "Missing Body Parameter(s)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Headers:
                  summary: "Missing Header Parameter(s)"
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Missing-Body:
                  summary: "Missing Body Parameter(s)"
                  value:
                    code: 10
                    message: "Missing Body Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        500:
          description: "Error sending email"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 0
                    message: "An unknown error occurred"
      security:
        - APIToken: []
...
  /phone/details:
    get:
      operationId: "GetPhoneDetails"
      description: |-
        Gets the details for the given phone number.
        Returns the details if successful, otherwise returns an error.

        Permissions Required: `Phone::Lookup`
      summary: "Get Phone Details"
      parameters:
        - in: "header"
          name: "x-number"
          description: "The phone number to get the details for"
          schema:
            type: string
          example: "+8888888888"
      responses:
        200:
          description: "The details for the given phone number"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumberData"
              examples:
                details:
                  value:
                    status: "success"
                    phone: "+8888888888"
                    phone_valid: true
                    phone_type: "fixed_line_or_mobile"
                    phone_region: "Michigan"
                    country: "United States"
                    country_code: "US"
                    country_prefix: "1"
                    international_number: "+1 888-888-888"
                    local_number: "(888) 888-8888"
                    e164: "+18888888888"
                    carrier: ""
        400:
          description: "Missing phone number information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Phone:
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-Phone:
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given phone number"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retrieve requested information"
      security:
        - APIToken: []
  /quotes/kanye:
    get:
      operationId: "GetKanyeQuote"
      description: |-
        Gets a random Kanye West quote.
        Returns the quote if successful, otherwise returns an error.

        Permissions Required: `Quote::Kanye`
      summary: "Get Kanye Quote"
      responses:
        200:
          description: "The Kanye West quote"
          content:
            plain/text:
              schema:
                type: string
              examples:
                quote:
                  value: "I feel like I'm too busy writing history to read it."
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
      security:
        - APIToken: []
  /quotes/ronswanson:
    get:
      operationId: "GetRonSwansonQuote"
      description: |-
        Gets a random Ron Swanson quote.
        Returns the quote if successful, otherwise returns an error.

        Permissions Required: `Quote::RonSwanson`
      summary: "Get Ron Swanson Quote"
      responses:
        200:
          description: "The Ron Swanson quote"
          content:
            plain/text:
              schema:
                type: string
              examples:
                quote:
                  value: "There's only one thing I hate more than lying: skim milk. Which is water that's lying about being milk."
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
      security:
        - APIToken: []
  /quotes/random:
    get:
      operationId: "GetRandomQuote"
      description: |-
        Gets a random quote.
        Returns the quote if successful, otherwise returns an error.

        Permissions Required: `Quote::Random`
      summary: "Get Random Quote"
      responses:
        200:
          description: "The random quote"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteData"
              examples:
                quote:
                  value:
                    quote: "The best way to predict the future is to create it."
                    author: "Abraham Lincoln"
                    tags:
                      - "future"
                      - "inspire"
                      - "life"
                      - "wisdom"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
      security:
        - APIToken: []
...
components:
  securitySchemes:
    APIToken:
      type: "http"
      scheme: "bearer"
      description: |-
        JSON Web Token Authentication with information containing your license key

        Permissions:
        - `Global::Data.Read` (0:0:0) - **Unused**
        - `Global::Data.Write` (0:0:1) - **Unused**
        - `Global::Data.Create` (0:0:2) - **Unused**
        - `Global::Data.Delete` (0:0:3) - **Unused**
        - `Global::Token.Read` (0:1:0) - Allows reading of token information
        - `Global::Token.ReadAll` (0:1:1) - Allows reading of all token information
        - `Global::Token.Write` (0:1:2) - **Unused**
        - `Global::Token.Create` (0:1:3) - Allows creating of token information
        - `Global::Token.Delete` (0:1:4) - Allows deleting of token 
        - `Global::User.Read` (0:2:0) - Read user information
        - `Global::User.ReadAll` (0:2:1) - **Unused**
        - `Global::User.Write` (0:2:2) - Write/Update user information
        - `Global::User.Create` (0:2:3) - Create a user profile
        - `Global::User.Delete` (0:2:4) - Delete a user profile
        - `Cryptography::Encrypt` (1:0) - Ability to encrypt data using the `/crypto/enc` endpoint
        - `Cryptography::Decrypt` (1:1) - Ability to decrypt data using the `/crypto/dec` endpoint
        - `Dictionary::Definition` (2:0) - !! DEPRECIATED!! Ability to get the definition of a word using the `/dict/def` endpoint
        - `Dictionary::Synonym` (2:1) - !! DEPRECIATED!! Ability to get the synonyms of a word using the `/dict/syn` endpoint
        - `Dictionary::Antonym` (2:2) - !! DEPRECIATED!! Ability to get the antonyms of a word using the `/dict/ant` endpoint
        - `Facts::Dog` (3:0) - Ability to get a random dog fact using the `/facts/dog` endpoint
        - `Facts::Cat` (3:1) - Ability to get a random cat fact using the `/facts/cat` endpoint
        - `Facts::ChuckNorris` (3:2) - Ability to get a random fact using the `/facts/chucknorris` endpoint
        - `Location::Coord` (4:0) - Ability to get the location of coordinates using the `/location/coords` endpoint
        - `Location::Pluscode` (4:1) - Ability to get the location of a pluscode using the `/location/pluscode` endpoint
        - `Location::Address` (4:2) - Ability to get the location of an address using the `/location/address` endpoint
        - `Phone::Lookup` (5:0) - Ability to get the details of a phone number using the `/phone/details` endpoint
        - `Phone::Validate` (5:1) - **Unused**
        - `Quote::Kanye` (6:0) - Ability to get a random Kanye West quote using the `/quotes/kanye` endpoint
        - `Quote::RonSwanson` (6:1) - Ability to get a random Ron Swanson quote using the `/quotes/ronswanson` endpoint
        - `Quote::Random` (6:2) - Ability to get a random quote using the `/quotes/random` endpoint
        - `Weather::Current` (7:0) - Ability to get the current weather using the `/weather/current` endpoint
        - `Minecraft::Hive.Player` (8:0:0) - Ability to get information about a player on the Hive Minecraft server using the `/minecraft/hive/player` endpoint
        - `Minecraft::Hive.Map` (8:0:1) - Ability to get information about a map on the Hive Minecraft server using the `/minecraft/hive/map` endpoint
...
  schemas:
    ErrorData:
      title: Error Response
      description: |-
        The error data that is returned when an error occurs

        All Error Codes are located in `/functions/errorResponse.js` [GitHub](https://github.com/femdevs/femdev-website/blob/main/functions/errorResponse.js)
      required:
        - code
        - message
      properties:
        code:
          nullable: false
          type: integer
          description: "The error code"
          minimum: 0
          maximum: 30
        message:
          nullable: false
          type: string
          description: "The error message"
...
    PhoneNumberData:
      title: Phone Number Response
      description: |-
        The information that is returned when getting phone number information
      required:
        - status
        - phone
        - phone_valid
        - phone_type
        - phone_region
        - country
        - country_code
        - country_prefix
        - international_number
        - local_number
        - e164
        - carrier
      properties:
        status:
          nullable: false
          type: string
          description: "The status of the phone number"
        phone:
          nullable: false
          type: string
          description: "The phone number"
        phone_valid:
          nullable: false
          type: boolean
          description: "Whether or not the phone number is valid"
        phone_type:
          nullable: false
          type: string
          description: "The type of the phone number"
        phone_region:
          nullable: false
          type: string
          description: "The region of the phone number"
        country:
          nullable: false
          type: string
          description: "The country of the phone number"
        country_code:
          nullable: false
          type: string
          description: "The country code of the phone number"
        country_prefix:
          nullable: false
          type: string
          description: "The country prefix of the phone number"
        international_number:
          nullable: false
          type: string
          description: "The international number of the phone number"
        local_number:
          nullable: false
          type: string
          description: "The local number of the phone number"
        e164:
          nullable: false
          type: string
          description: "The E164 number of the phone number"
        carrier:
          nullable: false
          type: string
          description: "The carrier of the phone number"
    QuoteData:
      title: Quote Response
      description: |-
        The information that is returned when getting a quote
      required:
        - quote
        - author
        - tags
      properties:
        quote:
          nullable: false
          type: string
          description: "The quote"
        author:
          nullable: false
          type: string
          description: "The author of the quote"
        tags:
          nullable: false
          type: array
          items:
            type: string
          description: "The tags of the quote"
    EmailReqText:
      title: Email Text
      description: |-
        The text contents of the email
      type: string
    EmailReqHTML:
      title: Email HTML
      description: |-
        The HTML contents of the email
      type: string
    EmailReqURL:
      title: Email URL
      description: |-
        The URL for the HTML contents of the email
      type: string
    EmailRequest:
      title: Email Request
      description: |-
        The information that is required to send an email
      required:
        - to
        - subject
      properties:
        to:
          nullable: false
          type: string
          description: "The email address to send the email to"
        subject:
          nullable: false
          type: string
          description: "The subject of the email"
      additionalProperties: 
        oneOf:
          - type: string
            $ref: "#/components/schemas/EmailReqText"
          - type: string
            $ref: "#/components/schemas/EmailReqHTML"
          - type: string
            $ref: "#/components/schemas/EmailReqURL"
...